first_period/trustid-components/trustid-internal-main/server/requirements.txt
celery==4.4.7
Django==2.2.17
django-cors-headers==3.10.1
djangorestframework==3.9.4
djangorestframework-simplejwt
dlib
drf-yasg==1.16.0
flower==0.9.5
gevent==21.8.0
gunicorn==19.9.0
imageio==2.13.3
joblib
jsonschema
matplotlib
opencv-python==4.1.0.25
psycopg2==2.8.6
pyjwt==1.7.1
torchvision==0.10
sklearn
graphviz
django-extensions
